Update group spec tests
Previously, a resource's GUID was determined by looking in
@property_hash for a value that was read by self.instances. This works
fine in the code because that value isn't checked by Puppet, but that's
hard to spec test. This commit does a lookup to dscl for the value -
which makes it easy to expect and test.

This also adds tests for the groups method.