Fix issue with setting groups in create method
Previously, in the create method, the GUID value was retrieved by
looking in the @property_hash. Because the GUID value is NOT in
the @property_hash when initially creating a new user, dscl
would fail.  This commit generates a new GUID with uuidgen, sets
the GUID for a new user, and uses it if the new user resource
contains a groups property value.