Set GID value correctly in create method
Previously, whatever value was passed for the GID property was
set as the user's PrimaryGroupID in the create method. If a
non-numerical value was passed (i.e. a group's name), that
value was literally set as the PrimaryGroupID (instead of the
group's GID number). The next time Puppet ran, the PrimaryGroupID
would be set to the GID of the group name passed. This commit
adds functionality to determine when a group name has been
passed and to determine (and set) its GID value on the first Puppet run.